Add beans creation to new config file
