+++ src/main/java/pl/cleankod/ApplicationInitializer.java
@@ -8,7 +8,9 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.SpringBootConfiguration;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
 import org.springframework.core.env.Environment;
+import pl.cleankod.exchange.core.config.ExchangeConfig;
 import pl.cleankod.exchange.core.gateway.AccountRepository;
 import pl.cleankod.exchange.core.gateway.CurrencyConversionService;
 import pl.cleankod.exchange.core.usecase.FindAccountAndConvertCurrencyUseCase;
@@ -21,56 +23,13 @@ import pl.cleankod.exchange.provider.nbp.ExchangeRatesNbpClient;
 
 import java.util.Currency;
 
-@SpringBootConfiguration
+
 @EnableAutoConfiguration
+@Import(ExchangeConfig.class)
 public class ApplicationInitializer {
     public static void main(String[] args) {
         SpringApplication.run(ApplicationInitializer.class, args);
     }
 
-    @Bean
-    AccountRepository accountRepository() {
-        return new AccountInMemoryRepository();
-    }
 
-    @Bean
-    ExchangeRatesNbpClient exchangeRatesNbpClient(Environment environment) {
-        String nbpApiBaseUrl = environment.getRequiredProperty("provider.nbp-api.base-url");
-        return Feign.builder()
-                .client(new ApacheHttpClient())
-                .encoder(new JacksonEncoder())
-                .decoder(new JacksonDecoder())
-                .target(ExchangeRatesNbpClient.class, nbpApiBaseUrl);
-    }
-
-    @Bean
-    CurrencyConversionService currencyConversionService(ExchangeRatesNbpClient exchangeRatesNbpClient) {
-        return new CurrencyConversionNbpService(exchangeRatesNbpClient);
-    }
-
-    @Bean
-    FindAccountUseCase findAccountUseCase(AccountRepository accountRepository) {
-        return new FindAccountUseCase(accountRepository);
-    }
-
-    @Bean
-    FindAccountAndConvertCurrencyUseCase findAccountAndConvertCurrencyUseCase(
-            AccountRepository accountRepository,
-            CurrencyConversionService currencyConversionService,
-            Environment environment
-    ) {
-        Currency baseCurrency = Currency.getInstance(environment.getRequiredProperty("app.base-currency"));
-        return new FindAccountAndConvertCurrencyUseCase(accountRepository, currencyConversionService, baseCurrency);
-    }
-
-    @Bean
-    AccountController accountController(FindAccountAndConvertCurrencyUseCase findAccountAndConvertCurrencyUseCase,
-                                        FindAccountUseCase findAccountUseCase) {
-        return new AccountController(findAccountAndConvertCurrencyUseCase, findAccountUseCase);
-    }
-
-    @Bean
-    ExceptionHandlerAdvice exceptionHandlerAdvice() {
-        return new ExceptionHandlerAdvice();
-    }
 }
