+++ src/main/java/pl/cleankod/exchange/core/config/ExchangeConfig.java
@@ -1,0 +1,70 @@
+package pl.cleankod.exchange.core.config;
+
+import feign.Feign;
+import feign.httpclient.ApacheHttpClient;
+import feign.jackson.JacksonDecoder;
+import feign.jackson.JacksonEncoder;
+import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.core.env.Environment;
+import pl.cleankod.exchange.core.gateway.AccountRepository;
+import pl.cleankod.exchange.core.gateway.CurrencyConversionService;
+import pl.cleankod.exchange.core.usecase.FindAccountAndConvertCurrencyUseCase;
+import pl.cleankod.exchange.core.usecase.FindAccountUseCase;
+import pl.cleankod.exchange.entrypoint.AccountController;
+import pl.cleankod.exchange.entrypoint.ExceptionHandlerAdvice;
+import pl.cleankod.exchange.provider.AccountInMemoryRepository;
+import pl.cleankod.exchange.provider.CurrencyConversionNbpService;
+import pl.cleankod.exchange.provider.nbp.ExchangeRatesNbpClient;
+
+import java.util.Currency;
+
+@SpringBootConfiguration
+public class ExchangeConfig {
+
+    @Bean
+    AccountRepository accountRepository() {
+        return new AccountInMemoryRepository();
+    }
+
+    @Bean
+    ExchangeRatesNbpClient exchangeRatesNbpClient(Environment environment) {
+        String nbpApiBaseUrl = environment.getRequiredProperty("provider.nbp-api.base-url");
+        return Feign.builder()
+                .client(new ApacheHttpClient())
+                .encoder(new JacksonEncoder())
+                .decoder(new JacksonDecoder())
+                .target(ExchangeRatesNbpClient.class, nbpApiBaseUrl);
+    }
+
+    @Bean
+    CurrencyConversionService currencyConversionService(ExchangeRatesNbpClient exchangeRatesNbpClient) {
+        return new CurrencyConversionNbpService(exchangeRatesNbpClient);
+    }
+
+    @Bean
+    FindAccountUseCase findAccountUseCase(AccountRepository accountRepository) {
+        return new FindAccountUseCase(accountRepository);
+    }
+
+    @Bean
+    FindAccountAndConvertCurrencyUseCase findAccountAndConvertCurrencyUseCase(
+            AccountRepository accountRepository,
+            CurrencyConversionService currencyConversionService,
+            Environment environment
+    ) {
+        Currency baseCurrency = Currency.getInstance(environment.getRequiredProperty("app.base-currency"));
+        return new FindAccountAndConvertCurrencyUseCase(accountRepository, currencyConversionService, baseCurrency);
+    }
+
+    @Bean
+    AccountController accountController(FindAccountAndConvertCurrencyUseCase findAccountAndConvertCurrencyUseCase,
+                                        FindAccountUseCase findAccountUseCase) {
+        return new AccountController(findAccountAndConvertCurrencyUseCase, findAccountUseCase);
+    }
+
+    @Bean
+    ExceptionHandlerAdvice exceptionHandlerAdvice() {
+        return new ExceptionHandlerAdvice();
+    }
+}
